Update to promotion endorsement table
Removing the Fellow engineer row as it does not give a concrete path to achieving the promotion.  We also don't know the list of endorsements so would just be randomly made up endorsements at this time.  As it is IMHO the endorsement could just create the appearance of being there to make a statement, even if a positive one.  So I think it's better to just remove it.